+++ HRInventories/HRInventories/Models/Reqestmaster.cs
@@ -10,7 +10,7 @@ namespace HRInventories.Models
             Requestdetail = new HashSet<Requestdetail>();
         }
 
-        public int Requestid { get; set; }
+        public long Requestid { get; set; }
         public string Employeeid { get; set; }
         public bool Isread { get; set; }
         public string Userid { get; set; }
