+++ HRInventories/HRInventories/Models/Requestdetail.cs
@@ -5,9 +5,9 @@ namespace HRInventories.Models
 {
     public partial class Requestdetail
     {
-        public int Requestdetailid { get; set; }
-        public int Requestid { get; set; }
-        public int Productid { get; set; }
+        public long Requestdetailid { get; set; }
+        public long Requestid { get; set; }
+        public long Productid { get; set; }
         public long Quantity { get; set; }
         public string Status { get; set; }
         public string Userid { get; set; }
