+++ HRInventories/HRInventories/UIModels/PodetailModel.cs
@@ -7,9 +7,9 @@ namespace HRInventories.UIModels
 {
     public class PodetailModel
     {
-        public int Podetailid { get; set; }
+        public long Podetailid { get; set; }
         public long Poid { get; set; }
-        public int Productid { get; set; }
+        public long Productid { get; set; }
         public long Quantity { get; set; }
         public double Porate { get; set; }
         public double Discount { get; set; }
